refactor: TMDb共有処理の責務分割
