+++ supabase/functions/_shared/tmdb/seasons.ts
@@ -1,0 +1,59 @@
+import { fetchTmdbJson, firstNonBlank } from "./http.ts";
+import {
+  type TmdbSearchResult,
+  type TmdbSeasonOption,
+  type TmdbTvDetailsResponse,
+} from "./types.ts";
+
+export async function fetchTmdbSeasonOptions(
+  result: TmdbSearchResult,
+): Promise<TmdbSeasonOption[]> {
+  if (result.tmdbMediaType !== "tv") {
+    return [];
+  }
+
+  const json = await fetchTmdbJson<TmdbTvDetailsResponse>(`/tv/${result.tmdbId}`, {
+    language: "ja-JP",
+  });
+
+  return (json.seasons ?? [])
+    .filter((season) => season.season_number > 1)
+    .map((season) => ({
+      seasonNumber: season.season_number,
+      title: resolveSeasonTitle(result.title, season.season_number, season.name?.trim()),
+      overview: season.overview ?? null,
+      posterPath: season.poster_path ?? null,
+      releaseDate: season.air_date ?? null,
+      episodeCount: typeof season.episode_count === "number" ? season.episode_count : null,
+    }));
+}
+
+export function resolveSeasonTitle(
+  seriesTitle: string,
+  seasonNumber: number,
+  ...candidates: Array<string | null | undefined>
+) {
+  const fallback = `${seriesTitle} シーズン${seasonNumber}`;
+  const chosen = firstNonBlank(...candidates);
+
+  if (!chosen || isGenericSeasonLabel(chosen)) {
+    return fallback;
+  }
+
+  if (chosen.toLowerCase().includes(seriesTitle.toLowerCase())) {
+    return chosen;
+  }
+
+  return `${seriesTitle} ${chosen}`;
+}
+
+function isGenericSeasonLabel(value: string) {
+  const normalized = value.trim().toLowerCase();
+
+  return (
+    /^season\s*\d+$/i.test(value.trim()) ||
+    /^シーズン\s*\d+$/.test(value.trim()) ||
+    normalized === "season" ||
+    normalized === "シーズン"
+  );
+}
